+++ Services/ProductService.cs
@@ -96,15 +96,6 @@ namespace CGullProject.Services
             return categories;
         }
 
-        public async Task<IEnumerable<Product>> GetAllProducts()
-        {
-            IEnumerable<Product> inventory =
-               await _context.Inventory.ToListAsync<Product>();
-
-         
-            return inventory;
-        }
-
         public async Task<IEnumerable<Product>> GetProductsbyCategory(int categoryId)
         {
             List<Product> products = new List<Product>();
@@ -225,24 +216,6 @@ namespace CGullProject.Services
             );
         }
 
-        public async Task<bool> ChangeSalesStatus(string itemId, bool status)
-        {
-            // search for the product in inventory
-            Product? product = await _context.Inventory.FindAsync(itemId);
-            if (product == null)
-            {
-                return false;
-            }
-
-            // change the status 
-            product.OnSale = status;
-
-            // save changes
-            await _context.SaveChangesAsync();
-
-            return true;
-        }
-
         public async Task<bool> RemoveProduct(string productId)
         {
             throw new NotImplementedException();
@@ -253,22 +226,5 @@ namespace CGullProject.Services
             throw new NotImplementedException();
         }
 
-        public async Task<IEnumerable<Product>> GetAllSalesProducts()
-
-        {
-            IEnumerable<Product> inventory =
-               await _context.Inventory.ToListAsync<Product>();
-
-            List<Product> toReturn = new List<Product>();
-            foreach (Product product in inventory)
-            {
-                if (product.OnSale)
-                {
-                    toReturn.Add(product);
-                }
-            }
-
-            return toReturn;
-        }
     }
 }
